@@ -1,71 +1,13 @@
-[build-system]
-requires = ["setuptools>=61.0", "wheel"]
-build-backend = "setuptools.build_meta"
-
 [project]
 name = "claude-agent-api-server"
-version = "1.0.0"
-description = "A client-server architecture for the Claude Agent SDK"
+version = "0.1.0"
+description = "Add your description here"
 readme = "README.md"
-requires-python = ">=3.8"
-license = {text = "MIT"}
-authors = [
-    {name = "Claude Agent API Server Contributors"}
-]
-keywords = ["claude", "api", "server", "agent", "sdk", "fastapi"]
-classifiers = [
-    "Development Status :: 4 - Beta",
-    "Intended Audience :: Developers",
-    "License :: OSI Approved :: MIT License",
-    "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.8",
-    "Programming Language :: Python :: 3.9",
-    "Programming Language :: Python :: 3.10",
-    "Programming Language :: Python :: 3.11",
-    "Programming Language :: Python :: 3.12",
-    "Framework :: FastAPI",
-    "Topic :: Software Development :: Libraries :: Python Modules",
-]
-
+requires-python = ">=3.12"
 dependencies = [
-    "fastapi>=0.104.0",
-    "uvicorn>=0.24.0",
-    "httpx>=0.25.0",
-    "pydantic>=2.0.0",
+    "claude-agent-sdk",
+    "fastapi>=0.119.1",
 ]
 
-[project.optional-dependencies]
-dev = [
-    "pytest>=7.0.0",
-    "pytest-asyncio>=0.21.0",
-    "black>=23.0.0",
-    "ruff>=0.1.0",
-    "mypy>=1.0.0",
-]
-
-[project.urls]
-Homepage = "https://github.com/yourusername/claude-agent-api-server"
-Documentation = "https://github.com/yourusername/claude-agent-api-server#readme"
-Repository = "https://github.com/yourusername/claude-agent-api-server"
-Issues = "https://github.com/yourusername/claude-agent-api-server/issues"
-
-[project.scripts]
-claude-agent-server = "server:main"
-claude-agent-client = "client:main"
-
-[tool.setuptools]
-packages = ["."]
-
-[tool.ruff]
-line-length = 100
-target-version = "py38"
-
-[tool.black]
-line-length = 100
-target-version = ["py38", "py39", "py310", "py311"]
-
-[tool.mypy]
-python_version = "3.8"
-warn_return_any = true
-warn_unused_configs = true
-disallow_untyped_defs = false
+[tool.uv.sources]
+claude-agent-sdk = { workspace = true }
